requirements.txt
Flask
pypdf
Pillow
google-genai
